Runtime: separate system host functions and tests into separate mod.
