+++ crates/runtime/src/logic/host_functions.rs
@@ -1,8 +1,10 @@
 
 mod blobs;
 mod storage;
+mod system;
 mod governance;
 
 pub use blobs::*;
 pub use storage::*;
+pub use system::*;
 pub use governance::*;
